+++ src/quantum_circuit_drawer/utils/__init__.py
@@ -1,5 +1,23 @@
 """Utility exports."""
 
-from .formatting import format_gate_name, format_parameter, format_parameters
+from .formatting import (
+    format_gate_name,
+    format_gate_name_mathtext,
+    format_parameter,
+    format_parameter_text,
+    format_parameter_text_mathtext,
+    format_parameters,
+    format_visible_label,
+    format_visible_label_mathtext,
+)
 
-__all__ = ["format_gate_name", "format_parameter", "format_parameters"]
+__all__ = [
+    "format_gate_name",
+    "format_gate_name_mathtext",
+    "format_parameter",
+    "format_parameter_text",
+    "format_parameter_text_mathtext",
+    "format_parameters",
+    "format_visible_label",
+    "format_visible_label_mathtext",
+]